Use double for expected (-e) sizes.
Ignore-this: d1498b8be582f35f8d0feb701cccface

darcs-hash:82abd7983664f22d78e5416e06b26c71da43e7c3

+++ src/Gnuplot.hs
@@ -9,7 +9,7 @@ import Control.Monad (when)
 import System.Directory (findExecutable)
 import Control.Concurrent (forkIO)
 
-gnuplot :: Num i => [String] -> [(String,[(Int,i)])] -> [Int] -> IO ()
+gnuplot :: Num i => [String] -> [(String,[(Int,i)])] -> [Double] -> IO ()
 gnuplot preamble cols hlines = do
   fe <- findExecutable "gnuplot"
   when (fe == Nothing) $ error "Couldn't find the 'gnuplot' executable - aborting"
